implemented sliders for changing height and width and size

+++ OverlayMod/Stat/IStat.cs
@@ -56,7 +56,7 @@ namespace OverlayMod.Stat.Stats
         {
             this.enabled = StatConfig.getConfigEntry<bool>(enumType, "enabled") ?? defaultEnabled;
             this.textObject.SetActive(this.enabled);
-            this.text.fontSize = (StatConfig.getConfigEntry<int>(enumType, "size") ?? defaultSize) * ((Plugin.scaleX + Plugin.scaleY) / 2); // last part averages the difference in text size incase the display ratio isnt 16:9
+            this.text.fontSize = (StatConfig.getConfigEntry<float>(enumType, "size") ?? defaultSize) * ((Plugin.scaleX + Plugin.scaleY) / 2); // last part averages the difference in text size incase the display ratio isnt 16:9
             this.text.alignment = optionalAllignmentOverride ?? TextAlignmentOptions.Center;
 
             float textPosX = (-Screen.width / 2) + ((StatConfig.getConfigEntry<float>(enumType, "posX") ?? defaultPosition.x) * Plugin.scaleX);
